Fix Hotkeys that are not modifiers or characters not working
When creating hotkeys, some keys were not working. Particularly those that are not modifiers or characters (e.g. Esc, Page Down, Arrows, End, etc...) They Key enums were getting added to the list of hotkeys by the parsing function instead of the KeyCode. I added logic to check if the key is not a modifier, and if its not, create and add the KeyCode to the hotkey instead.