Do not block when join has timeout
When calling join on a listener, any arguments to the method are passed
to Thread.join. A timeout argument indicates that the user does not want
the method to block.

After joining the listener thread, any exceptions raised in callbacks
are read from a queue and rethrown. However, if the thread join
terminated because of a timeout, the queue will be empty, and the read
operation will block.

This commit makes an attempt at respecting the original timeout.